+++ changelog.md
@@ -1,3 +1,7 @@
+## 15-08-2024
+
+- Added board and tested.
+
 ## 14-08-2024
 
 - Added basic piece, pawn and rook class. Basic piece tests.